Fix build break
-Build was breaking on Linux due to conflict between uint64_t and unsigned long long. Removed
use of unsigned long long and use uint64_t from now on for 64-bit values.

+++ KeyFinder/main.cpp
@@ -65,9 +65,9 @@ void statusCallback(KeyFinderStatusInfo info)
 
 	std::string timeStr = "[" + util::formatSeconds((unsigned int)(info.totalTime / 1000)) + "]";
 
-	std::string usedMemStr = util::format((info.deviceMemory - info.freeMemory) / (unsigned long long)(1024 * 1024));
+	std::string usedMemStr = util::format((info.deviceMemory - info.freeMemory) /(1024 * 1024));
 
-	std::string totalMemStr = util::format(info.deviceMemory / (unsigned long long)(1024 * 1024));
+	std::string totalMemStr = util::format(info.deviceMemory / (1024 * 1024));
 
 	std::string targetStr = util::format(info.targets) + " target";
 	if(info.targets > 1) {
